+++ src/Filter/Common/EnumFilter.php
@@ -33,6 +33,9 @@ class EnumFilter implements FilterInterface
 
         $alias = $qb->getRootAliases()[0];
         $column = $context->getField()->getOriginName();
+        if ($context->getField()->getOriginType() === 'ReflectionMethod') {
+            $column = $context->getField()->getName();
+        }
         switch ($condition->getOp()) {
             case NodeComparisonOperatorType::IN:
                 $qb->andWhere($qb->expr()->in("{$alias}.{$column}", $condition->getValues()));