+++ src/Filter/Common/NodeFilter.php
@@ -39,6 +39,9 @@ class NodeFilter implements FilterInterface
         $entity = $context->getNode()->getClass();
         $metadata = $qb->getEntityManager()->getClassMetadata($entity);
         $column = $context->getField()->getOriginName();
+        if ($context->getField()->getOriginType() === 'ReflectionMethod') {
+            $column = $context->getField()->getName();
+        }
 
         if (!$metadata->hasAssociation($column)) {
             throw new \RuntimeException(sprintf('There are not valid association in %s called %s.', $entity, $column));