+++ CHANGELOG.md
@@ -4,6 +4,7 @@ v1.3.5 - Unreleased
  * Fix bug with redis pub/sub handler and socket timeouts
  * Fix bug when use search in node pagination with  chars like `'%_`
  * Add support to use native javascript date objects as input for dates
+ * Fix bug with list filters for interface fields
 
 v1.3.4 - 2019-05-03
 ----